remove extractLocation class move to func
